chore(ssr): mostly revert move to @lwc/shared
